[Chore] #66 - member와 관련된 패키지 이동
