+++ src/main/java/com/arabook/arabook/member/entity/AIRecommendation.java
@@ -1,4 +1,6 @@
-package com.arabook.arabook.storage.domain.member.entity;
+package com.arabook.arabook.member.entity;
+
+import java.time.LocalDate;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,34 +10,30 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotNull;
 
-import com.arabook.arabook.category.entity.SubCategory;
+import com.arabook.arabook.book.entity.Book;
 
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "member_sub_category_selections")
-public class MemberSubCategorySelection {
+@Table(name = "daily_book_id_recommendations")
+public class AIRecommendation {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long memberSubCategorySelectionId;
+  private Long dailyBookIdRecommendationId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "member_id")
   private Member member;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "category_id")
-  private SubCategory subCategory;
+  @JoinColumn(name = "book_id")
+  private Book book;
 
-  @Builder
-  private MemberSubCategorySelection(Member member, SubCategory subCategory) {
-    this.member = member;
-    this.subCategory = subCategory;
-  }
+  @NotNull LocalDate recommendationDate;
 }
