+++ src/main/java/com/arabook/arabook/member/entity/enums/Gender.java
@@ -1,0 +1,7 @@
+package com.arabook.arabook.member.entity.enums;
+
+public enum Gender {
+  MAN,
+  WOMAN,
+  UNKNOWN
+}
