+++ src/main/java/com/arabook/arabook/member/entity/enums/SocialPlatformType.java
@@ -1,0 +1,6 @@
+package com.arabook.arabook.member.entity.enums;
+
+public enum SocialPlatformType {
+  KAKAO,
+  APPLE
+}
