+++ src/main/java/com/arabook/arabook/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.storage.domain.member.repository;
+package com.arabook.arabook.member.repository;
 
 import static com.arabook.arabook.common.exception.member.MemberExceptionType.*;
 
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.arabook.arabook.common.exception.member.MemberException;
-import com.arabook.arabook.storage.domain.member.entity.Member;
+import com.arabook.arabook.member.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
   Optional<Member> findByMemberId(final Long memberId);
