+++ src/main/java/com/arabook/arabook/member/repository/MemberSubCategorySelectionRepository.java
@@ -1,9 +1,9 @@
-package com.arabook.arabook.storage.domain.member.repository;
+package com.arabook.arabook.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.arabook.arabook.storage.domain.member.entity.Member;
-import com.arabook.arabook.storage.domain.member.entity.MemberSubCategorySelection;
+import com.arabook.arabook.member.entity.Member;
+import com.arabook.arabook.member.entity.MemberSubCategorySelection;
 
 public interface MemberSubCategorySelectionRepository
     extends JpaRepository<MemberSubCategorySelection, Long> {
