+++ src/main/java/com/arabook/arabook/member/service/MemberService.java
@@ -1,0 +1,7 @@
+package com.arabook.arabook.member.service;
+
+import com.arabook.arabook.member.controller.dto.request.MemberOnboardingRequest;
+
+public interface MemberService {
+  void onboarding(MemberOnboardingRequest request, Long memberId);
+}
