+++ src/main/java/com/arabook/arabook/member/service/MemberSubCategorySelectionService.java
@@ -1,8 +1,8 @@
-package com.arabook.arabook.api.member.service;
+package com.arabook.arabook.member.service;
 
 import java.util.List;
 
-import com.arabook.arabook.storage.domain.member.entity.Member;
+import com.arabook.arabook.member.entity.Member;
 
 public interface MemberSubCategorySelectionService {
   void selectSubCategories(Member member, List<Long> categoryId);
