+++ src/main/java/com/arabook/arabook/member/service/MemberSubCategorySelectionServiceImpl.java
@@ -1,4 +1,4 @@
-package com.arabook.arabook.api.member.service;
+package com.arabook.arabook.member.service;
 
 import static com.arabook.arabook.common.exception.category.CategoryExceptionType.*;
 
@@ -11,9 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.arabook.arabook.category.entity.SubCategory;
 import com.arabook.arabook.category.repository.SubCategoryRepository;
 import com.arabook.arabook.common.exception.category.CategoryException;
-import com.arabook.arabook.storage.domain.member.entity.Member;
-import com.arabook.arabook.storage.domain.member.entity.MemberSubCategorySelection;
-import com.arabook.arabook.storage.domain.member.repository.MemberSubCategorySelectionRepository;
+import com.arabook.arabook.member.entity.Member;
+import com.arabook.arabook.member.entity.MemberSubCategorySelection;
+import com.arabook.arabook.member.repository.MemberSubCategorySelectionRepository;
 
 import lombok.RequiredArgsConstructor;
