--- src/main/java/com/arabook/arabook/storage/domain/member/entity/enums/Gender.java
@@ -1,7 +1,0 @@
-package com.arabook.arabook.storage.domain.member.entity.enums;
-
-public enum Gender {
-  MAN,
-  WOMAN,
-  UNKNOWN
-}
