--- src/main/java/com/arabook/arabook/storage/domain/member/entity/enums/SocialPlatformType.java
@@ -1,6 +1,0 @@
-package com.arabook.arabook.storage.domain.member.entity.enums;
-
-public enum SocialPlatformType {
-  KAKAO,
-  APPLE
-}
