+++ src/main/java/com/arabook/arabook/storage/domain/review/entity/Review.java
@@ -18,8 +18,8 @@ import jakarta.validation.constraints.NotNull;
 
 import com.arabook.arabook.book.entity.Book;
 import com.arabook.arabook.common.exception.review.ReviewException;
+import com.arabook.arabook.member.entity.Member;
 import com.arabook.arabook.storage.domain.common.entity.BaseTimeEntity;
-import com.arabook.arabook.storage.domain.member.entity.Member;
 import com.arabook.arabook.storage.domain.review.entity.enums.ReviewTag;
 
 import lombok.AccessLevel;
